Add test for response_model_by_alias working with openapi models
